Refactor actions into modules
