+++ .gitignore
@@ -1,4 +1,5 @@
 node_modules/
 config.js
+!/src/config.js
 /data/
 *.log
