+++ src/index.js
@@ -1,15 +1,14 @@
 const { program } = require('commander');
-const winMan = require('./actions');
+const winMan = require('./lib');
 
 start();
 
 async function start() {
-  program
-    .option('--first')
+  program.option('--first');
 
-  program.command('place').action(winMan.placeWindows)
-  program.command('store').action(winMan.storeWindows)
-  program.command('restore').action(winMan.restoreWindows)
+  program.command('place').action(winMan.placeWindows);
+  program.command('store').action(winMan.storeWindows);
+  program.command('restore').action(winMan.restoreWindows);
 
   program.command('stats').action(() => {
     const stats = winMan.getStats();
@@ -17,6 +16,4 @@ async function start() {
   });
 
   program.parse();
-  // const options = program.opts();
 }
-module.exports = {}
