+++ src/lib/index.js
@@ -1,0 +1,11 @@
+module.exports = {
+  ...require('../placement'),
+  ...require('../windows'),
+  ...require('../monitors'),
+  ...require('../fancyzones'),
+  ...require('../virtual-desktop'),
+  ...require('../store'),
+  ...require('../stats'),
+  ...require('../wallpapers'),
+  ...require('../config'),
+};
